Refactor GameFixes
- Add documentation comments for all patches
- Convert all hooks to Harmony patches and remove MMHOOK dependency
- Integrate the majority of fixes in ErrorPolice (delete GameFixes.cs)
- Put server method calls on client and asset fixes in separate files

These changes call for a static class for logging, so it can be accessed
from anywhere more conveniently than from the plugin class.

TODO: For some reason this breaks the RuleChoiceController.FindNetworkUser
fix. The set of instructions that are passed on to the patcher do not
correspond with what ILSpy says they should be. No idea why this difference
from MMHOOK to Harmony, but I'm commiting this as is and I'll investigate
later. Worse case scenario I modify the matched instructions according to
what the patcher method sees.