Fix the Frolic Teleport fix failing when only 1 node is available
For some reason the random list index is selected from [1, list.Count)
instead from [0, list.Count), so we need to make sure that list.Count > 1.

We're also changing the early return check to just encode the necessary
information in `flag` and `vector` and letting the method play out as
normal for the teleport since the TP effect has already been fired before
the patched method is called.