Fix some NRE edge cases for the VineOrb fix
Turns out null checking `self.target.healthComponent.body` can return true
for a bunch of Jellyfish killed by gasoline, yet `self.target` is null when
trying to access it for the sound playing.

Harmony internals require there be only a single Ret instruction, for which
it can provide a label to, so use this to "skip" the whole method instead
of emitting our own Ret.